Initial Development
This is the initial development before i remembered github existed, comments in code describe in detail but in a summary: The program so far utilized an object class holding important info about the item and generating a ID number for easy tracking. So far it utilizes a dictionary to remember all the items but i might change into a csv file so that the program does not have to constantly keep running the same instance for the checkout part to work in interdependence. Meaning creating one big file holding all the items everytime the processing part of the project is run. and whenever the checkout part of the code is ran we can just feed the big file.